+++ azure/functions/decorators/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
-from ._http import HttpMethod
+from .http import HttpMethod
 from .core import Cardinality, AccessRights
 from .function_app import FunctionApp, Function, DataType, AuthLevel
