+++ tests/decorators/test_decorators.py
@@ -2,7 +2,7 @@
 #  Licensed under the MIT License.
 import unittest
 
-from azure.functions.decorators._http import HttpTrigger, HttpMethod
+from azure.functions.decorators.http import HttpTrigger, HttpMethod
 from azure.functions.decorators.constants import TIMER_TRIGGER, HTTP_TRIGGER, \
     HTTP_OUTPUT, QUEUE, QUEUE_TRIGGER, SERVICE_BUS, SERVICE_BUS_TRIGGER, \
     EVENT_HUB, EVENT_HUB_TRIGGER, COSMOS_DB, COSMOS_DB_TRIGGER
