+++ tests/decorators/test_function_app.py
@@ -4,8 +4,9 @@ import unittest
 from unittest import mock
 
 from azure.functions import WsgiMiddleware, AsgiMiddleware
-from azure.functions.decorators._http import HttpTrigger, HttpOutput, \
+from azure.functions.decorators.http import HttpTrigger, HttpOutput, \
     HttpMethod
+from azure.functions.decorators.blob import BlobTrigger
 from azure.functions.decorators.constants import HTTP_OUTPUT, HTTP_TRIGGER
 from azure.functions.decorators.core import DataType, AuthLevel, \
     BindingDirection, SCRIPT_FILE_NAME
